+++ CHANGELOG.md
@@ -14,6 +14,22 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
+## [29.85.3] - 2026-02-13
+Move needScatterGather method to Client interface
+
+## [29.85.2] - 2026-02-13
+Adding logging in D2 client for ZK usages
+
+## [29.85.1] - 2026-02-12
+Add public needScatterGather method to RestClient
+
+## [29.85.0] - 2026-02-11
+Add getClusterName method to D2Client
+
+## [29.84.1] - 2026-02-10
+- Add missing ClusterProprties fields load balancing policies and outlier detection.
+- Use aliased union for load balancing policy config.
+
 ## [29.84.0] - 2026-02-03
 - Add load balancing policies to D2Cluster and LB config for RingHash, LeastRequest, and RoundRobin.
 
@@ -5948,7 +5964,12 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.84.0...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.85.3...master
+[29.85.3]: https://github.com/linkedin/rest.li/compare/v29.85.2...v29.85.3
+[29.85.2]: https://github.com/linkedin/rest.li/compare/v29.85.1...v29.85.2
+[29.85.1]: https://github.com/linkedin/rest.li/compare/v29.85.0...v29.85.1
+[29.85.0]: https://github.com/linkedin/rest.li/compare/v29.84.1...v29.85.0
+[29.84.1]: https://github.com/linkedin/rest.li/compare/v29.84.0...v29.84.1
 [29.84.0]: https://github.com/linkedin/rest.li/compare/v29.83.1...v29.84.0
 [29.83.1]: https://github.com/linkedin/rest.li/compare/v29.83.0...v29.83.1
 [29.83.0]: https://github.com/linkedin/rest.li/compare/v29.82.0...v29.83.0